pyproject.toml
[tool.poetry]
name = "python-quickstart"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "app"},
]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
fastapi = "^0.115.4"
pytest = "^8.3.3"
requests = "^2.32.3"
pytz = "^2024.2"
toml = "^0.10.2"
python-dotenv = "^1.0.1"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
pytest-order = "^1.3.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
black = "^24.10.0"
pydantic = "^2.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
